pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llama-rag-kb"
version = "0.1.0"
description = "RAG Knowledge Base using Llama.cpp and GGUF models"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "llama-cpp-python",
    "langchain-chroma",
    "langchain-text-splitters",
    "langchain-core",
    "langchain-community",
    "gradio",
    "gradio-pdf",
    "pypdf>=6.1.3",
]
requires-python = ">=3.12"

[project.scripts]
llama-rag-kb = "app:main"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v"
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    ".venv/*",
    "*/site-packages/*"
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
